application/controllers.py
from flask import request, url_for, redirect, jsonify
from flask_login import login_user, logout_user, current_user, login_required
from flask_security import auth_required, login_user, hash_password, roles_required
from flask import render_template
from flask import current_app as app
from werkzeug.utils import secure_filename
from application.database import db
from datetime import datetime
from datetime import timedelta
import os
from email_validator import validate_email
from application.models import Users, Requests, Sections, Books, Books_Authors, Books_Lendings, Authors
import jwt

@app.route("/")
def index():
    return render_template("index.html")
