requirements.txt
Flask
Gunicorn
flask_security
werkzeug
datetime
os
email_validator
jwt
flask_restful
flask_cors
celery
flask_sqlalchemy
flask_caching
smtplib
email
enum
flask_excel
jinja2